Move pullrequest specific logic to the PullRequestList component
