+++ app/src/ui/branches/pull-request-list.tsx
@@ -11,7 +11,14 @@ import { PullRequest } from '../../models/pull-request'
 import { NoPullRequests } from './no-pull-requests'
 import { IMatches } from '../../lib/fuzzy-find'
 import { Dispatcher } from '../dispatcher'
-import { GitHubRepository } from '../../models/github-repository'
+import {
+  RepositoryWithGitHubRepository,
+  getNonForkGitHubRepository,
+} from '../../models/repository'
+import { Button } from '../lib/button'
+import { Octicon, OcticonSymbol } from '../octicons'
+import { FoldoutType } from '../../lib/app-state'
+import { startTimer } from '../lib/timing'
 
 interface IPullRequestListItem extends IFilterListItem {
   readonly id: string
@@ -28,35 +35,17 @@ interface IPullRequestListProps {
   /** The currently selected pull request */
   readonly selectedPullRequest: PullRequest | null
 
-  /** The name of the repository. */
-  readonly repositoryName: string
-
   /** Is the default branch currently checked out? */
   readonly isOnDefaultBranch: boolean
-
-  /** The current filter text to render */
-  readonly filterText: string
-
-  /** Called when the user clicks on a pull request. */
-  readonly onItemClick: (pullRequest: PullRequest) => void
 
   /** Called when the user wants to dismiss the foldout. */
   readonly onDismiss: () => void
 
-  /** Callback to fire when the filter text is changed */
-  readonly onFilterTextChanged: (filterText: string) => void
-
   /** Called when the user opts to create a branch */
   readonly onCreateBranch: () => void
 
-  /** Called when the user opts to create a pull request */
-  readonly onCreatePullRequest: () => void
-
-  /** Called to render content after the filter. */
-  readonly renderPostFilter?: () => JSX.Element | null
-
   /** Callback fired when user selects a new pull request */
-  readonly onSelectionChanged?: (
+  readonly onSelectionChanged: (
     pullRequest: PullRequest | null,
     source: SelectionSource
   ) => void
@@ -70,15 +59,14 @@ interface IPullRequestListProps {
   ) => void
 
   readonly dispatcher: Dispatcher
-  readonly repository: GitHubRepository
+  readonly repository: RepositoryWithGitHubRepository
 
   /** Are we currently loading pull requests? */
   readonly isLoadingPullRequests: boolean
-
-  readonly renderListHeader?: () => JSX.Element | null
 }
 
 interface IPullRequestListState {
+  readonly filterText: string
   readonly groupedItems: ReadonlyArray<IFilterListGroup<IPullRequestListItem>>
   readonly selectedItem: IPullRequestListItem | null
 }
@@ -116,6 +104,7 @@ export class PullRequestList extends React.Component<
     const selectedItem = resolveSelectedItem(group, props, null)
 
     this.state = {
+      filterText: '',
       groupedItems: [group],
       selectedItem,
     }
@@ -139,15 +128,15 @@ export class PullRequestList extends React.Component<
         groups={this.state.groupedItems}
         selectedItem={this.state.selectedItem}
         renderItem={this.renderPullRequest}
-        filterText={this.props.filterText}
-        onFilterTextChanged={this.props.onFilterTextChanged}
+        filterText={this.state.filterText}
+        onFilterTextChanged={this.onFilterTextChanged}
         invalidationProps={this.props.pullRequests}
         onItemClick={this.onItemClick}
         onSelectionChanged={this.onSelectionChanged}
         onFilterKeyDown={this.props.onFilterKeyDown}
-        renderGroupHeader={this.props.renderListHeader}
+        renderGroupHeader={this.renderListHeader}
         renderNoItems={this.renderNoItems}
-        renderPostFilter={this.props.renderPostFilter}
+        renderPostFilter={this.renderPostFilter}
       />
     )
   }
@@ -155,12 +144,12 @@ export class PullRequestList extends React.Component<
   private renderNoItems = () => {
     return (
       <NoPullRequests
-        isSearch={this.props.filterText.length > 0}
+        isSearch={this.state.filterText.length > 0}
         isLoadingPullRequests={this.props.isLoadingPullRequests}
-        repositoryName={this.props.repositoryName}
+        repositoryName={this.getRepositoryName()}
         isOnDefaultBranch={this.props.isOnDefaultBranch}
         onCreateBranch={this.props.onCreateBranch}
-        onCreatePullRequest={this.props.onCreatePullRequest}
+        onCreatePullRequest={this.onCreatePullRequest}
       />
     )
   }
@@ -184,22 +173,72 @@ export class PullRequestList extends React.Component<
     )
   }
 
-  private onItemClick = (item: IPullRequestListItem) => {
-    if (this.props.onItemClick) {
-      this.props.onItemClick(item.pullRequest)
-    }
+  private onItemClick = (
+    item: IPullRequestListItem,
+    source: SelectionSource
+  ) => {
+    const pullRequest = item.pullRequest
+
+    this.props.dispatcher.closeFoldout(FoldoutType.Branch)
+    const timer = startTimer(
+      'checkout pull request from list',
+      this.props.repository
+    )
+    this.props.dispatcher
+      .checkoutPullRequest(this.props.repository, pullRequest)
+      .then(() => timer.done())
+
+    this.props.onSelectionChanged(pullRequest, source)
   }
 
   private onSelectionChanged = (
     selectedItem: IPullRequestListItem | null,
     source: SelectionSource
   ) => {
-    if (this.props.onSelectionChanged) {
-      this.props.onSelectionChanged(
-        selectedItem != null ? selectedItem.pullRequest : null,
-        source
-      )
-    }
+    this.props.onSelectionChanged(
+      selectedItem != null ? selectedItem.pullRequest : null,
+      source
+    )
+  }
+
+  private getRepositoryName(): string {
+    return getNonForkGitHubRepository(this.props.repository).fullName
+  }
+
+  private renderListHeader = () => {
+    return (
+      <div className="filter-list-group-header">
+        Pull requests in {this.getRepositoryName()}
+      </div>
+    )
+  }
+
+  private onRefreshPullRequests = () => {
+    this.props.dispatcher.refreshPullRequests(this.props.repository)
+  }
+
+  private renderPostFilter = () => {
+    return (
+      <Button
+        disabled={this.props.isLoadingPullRequests}
+        onClick={this.onRefreshPullRequests}
+        tooltip="Refresh the list of pull requests"
+      >
+        <Octicon
+          symbol={OcticonSymbol.sync}
+          className={this.props.isLoadingPullRequests ? 'spin' : undefined}
+        />
+      </Button>
+    )
+  }
+
+  private onFilterTextChanged = (text: string) => {
+    this.setState({ filterText: text })
+  }
+
+  private onCreatePullRequest = () => {
+    this.props.dispatcher.closeFoldout(FoldoutType.Branch)
+    this.props.dispatcher.createPullRequest(this.props.repository)
   }
 }
